requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
click==8.3.2
fastapi==0.135.3
h11==0.16.0
idna==3.11
nulltype==2.3.1
plaid-python==39.0.0
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
six==1.17.0
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.44.0
PyJWT==2.9.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
requests==2.32.3
cachetools==5.5.2
